+++ packages/utils/parcel.d.ts
@@ -1,6 +1,6 @@
 // Parcel doesn’t seem to understand tsconfig’s plugin: typescript-plugin-css-modules
 // so we declare this and reference from each package’s src/index.ts (these are Parcel’s entry files via `source`)
-// this: /// <reference types="@rms/forge-utils/parcel" />
+// this: /// <reference types="@rishim777/forge-utils/parcel" />
 declare module '*.module.scss' {
   const value: Record<string, string>;
   export default value;