+++ packages/text/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @rms/forge-text
+# @rishim777/forge-text
 
 ## 2.0.0
 
@@ -11,4 +11,4 @@
 
 - Updated dependencies
   [[`fb5bc53`](https://github.com/rishim777/RMS-FORGE/commit/fb5bc53f82ea63888d497db5153e4eb0ffe6006f)]:
-  - @rms/forge-utils@2.0.0
+  - @rishim777/forge-utils@2.0.0